feat(simulator): money metric series become Decimal end to end
The metric series total_debt and clearing_volume are money by the
owner's own domain document, and they travelled as float from the
moment they were read out of the database to the moment they reached the
wire. They now stay Decimal the whole way, land in a Numeric(20,8)
column matching how the domain already stores money, and are rendered as
plain decimal strings.

Old rows are archived rather than converted. Converting a lossy float
into a decimal would have produced exact-looking digits that were never
measured, which is precisely what the owner forbade; the recorded
history's precision is gone and pretending otherwise would launder it.
The live table therefore starts empty, which is correct for simulator
telemetry that is disposable by construction.

Internal adversarial review found three P1 defects in the first two
rounds of this slice, and all three are fixed here.

The archive itself was lossy. It wrote the float through CAST AS TEXT,
whose digits depend on the extra_float_digits setting the migration
never pinned - measured, 0.30000000000000004 archived as "0.3" - while
the live rows were deleted in the same transaction, leaving the archive
as the only copy. The setting is now pinned, and more importantly the
migration gained a gate: every non-null live value must have an archived
string that casts back to the identical float, or it raises and the
transaction rolls back with the live rows untouched. A migration that
exists to refuse laundering must not be able to launder by accident.

The SQLite path still narrows money through binary64, and the change
made that worse rather than better: a lossy value was now being
presented as an exact decimal string. SQLite is not forbidden - it is
the repository default and the documented way to run locally, and
banning it would be a product decision - so the limitation is declared
instead: a warning once per run when a money series is actually
persisted, documented in both the storage and the wire references, and
pinned by a test that asserts the limitation as a limitation and says
what to delete if it ever disappears.

The third defect was created by this slice. A Numeric column can
overflow where float8 could not, and the writer answered failure by
rolling back the caller's tick session and swallowing the exception, so
the whole batch vanished, anything the tick had staged went with it, and
carry-forward then drew the last good measurement as a live flat line -
a permanently broken writer looking healthy, the very class this program
exists to remove. The write now runs in a savepoint, so only its own
work is undone. This is a sibling of B-A2b-001, which stays with the
triage that owns it: the same pattern in the bottlenecks writer is
recorded, not quietly absorbed, because this slice did not change its
reachability.

Both producers now emit one canonical form at the column's own scale,
because the same endpoint was returning "45.3" from one branch and
"45.30000000" from the other. Migration 018 also gained real coverage:
the previous test proved the model through create_all, not the revision,
so the new one executes upgrade and downgrade against a disposable
schema and asserts the archive survives a hostile extra_float_digits.

Gates repeated by the orchestrator: ruff clean, Alembic single head,
unit tier 1034 passed and 2 skipped, contract 23 passed, and the
Postgres marker tier 47 passed - up from 40 when this program started.
The column type was read back from information_schema as numeric(20,8)
rather than trusted from the model.

+++ app/db/models/simulator_storage.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
-from sqlalchemy import CheckConstraint, DateTime, Float, Index, Integer, JSON, String, func
+from decimal import Decimal
+
+from sqlalchemy import CheckConstraint, DateTime, Float, Index, Integer, JSON, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -70,7 +72,13 @@ class SimulatorRunMetric(Base):
     key: Mapped[str] = mapped_column(String(50), primary_key=True)
     t_ms: Mapped[int] = mapped_column(Integer, primary_key=True)
 
-    value: Mapped[float | None] = mapped_column(Float, nullable=True)
+    # 2026-08-20 / p007_t715: the money-carrying series (`total_debt`,
+    # `clearing_volume`) are declared as amounts by the domain model, so this
+    # column stores exact decimals - the same Numeric(20, 8) the debt and
+    # trustline tables use. The non-money series share the column and become
+    # exact too. NULL still means "not measured" and stays distinguishable from
+    # a measured zero (spec 007, F-007-1).
+    value: Mapped[Decimal | None] = mapped_column(Numeric(20, 8), nullable=True)
 
     __table_args__ = (
         CheckConstraint("t_ms >= 0", name="chk_simulator_run_metrics_t_ms"),
